move fizzbuzz.dart to src/
